final polish pre–submission
- add AI_Reflection.txt (explicit AI usage disclosure)
- README: tighten db setup, scope notes
- add-match UX: no redirect → stay on home + refresh table
- unify & normalize english validation messages
- DtoMapper / VenueDto small cleanups
- introduce NotFoundException + map to 404 (separate from domain 400)
- update services + controller tests accordingly
- minor MatchService signature/test alignment

+++ backend/src/main/java/com/sportradar/inesoitzinger/mappers/DtoMapper.java
@@ -36,7 +36,7 @@ public class DtoMapper {
                 e.getId(),
                 e.getName(),
                 e.getCity(),
-                Math.toIntExact(e.getCapacity())
+                (long) Math.toIntExact(e.getCapacity())
         );
     }
